requirements.txt
streamlit
pandas
plotly
gspread
oauth2client
pypdf
requests
google-generativeai